requirements.txt
websocket-client==1.6.3
requests==2.31.0
rel==0.4.9